Fix problems with responses and change structure of DTOs and requests

+++ src/main/kotlin/com/alejandro/library/payloads/serverresponse/AuthorResponse.kt
@@ -5,7 +5,7 @@ import com.alejandro.library.payloads.dto.AuthorDTO
 class AuthorResponse(
     override val next: String?,
     override val previous: String?,
-    override val count: Int,
+    override val count: Long,
     override val result: List<AuthorDTO>?,
     override val success: Boolean,
     override val error: Any?
